+++ src/display.rs
@@ -158,6 +158,11 @@ display_fn! {
     display_uri_component => uri::write_uri_component
 }
 
+display_fn! {
+    /// zero-allocation display wrapper for [`for_uri_path`](crate::for_uri_path).
+    display_uri_path => uri::write_uri_path
+}
+
 // -- json --
 
 display_fn! {
@@ -347,6 +352,7 @@ mod tests {
         display_uri_component,
         crate::for_uri_component
     );
+    display_matches_for!(uri_path, display_uri_path, crate::for_uri_path);
 
     // json
     display_matches_for!(json, display_json, crate::for_json);